+++ src/mons/inutia/ChainExpansion.sol
@@ -72,7 +72,7 @@ contract ChainExpansion is IMoveSet, BasicEffect {
      *  Effect implementation
      */
     // Steps: OnMonSwitchIn
-    function getStepsBitmap() external pure override returns (uint16) {
+    function getStepsBitmap() external pure override returns (uint32) {
         return 0x8010;
     }
