+++ src/mons/nirvamma/HardReset.sol
@@ -83,7 +83,7 @@ contract HardReset is IMoveSet, BasicEffect {
     }
 
     // Steps: AfterMove, ALWAYS_APPLIES (0x8000)
-    function getStepsBitmap() external pure override returns (uint16) {
+    function getStepsBitmap() external pure override returns (uint32) {
         return 0x8080;
     }
