+++ test/mocks/DummyStatus.sol
@@ -11,7 +11,7 @@ contract DummyStatus is BasicEffect {
     }
 
     // Steps: None
-    function getStepsBitmap() external pure override returns (uint16) {
+    function getStepsBitmap() external pure override returns (uint32) {
         return 0x00;
     }
 }
